barcode: Adding generic barcode scanner
Generic barcode scanner fragment is added. It can scan any barcode and
pass the value to parent activity. This scanner should be used in whole app.

Example of how to use this scanner is provided in
barcode/TestBarcodeScannerActivity  activity.

Any Activity who wants to use barcode scanner should implement IQRResult
interface. It is been demonstrated in that exmaple code

Signed-off-by: patilsan90 <patil.san90@gmail.com>